+++ Library/Subscription.cs
@@ -11,7 +11,7 @@ namespace Recurly
     /// </summary>
     public class Subscription : RecurlyEntity
     {
-        // changed to flags based on http://docs.recurly.com/api/subscriptions saying Subscriptions can be in multiple states
+        // changed to flags based on https://dev.recurly.com/docs/list-subscriptions saying Subscriptions can be in multiple states
         [Flags]
         // The currently valid Subscription States
         public enum SubscriptionState : short
@@ -82,6 +82,8 @@ namespace Recurly
         public string Currency { get; set; }
         public int Quantity { get; set; }
 
+        public bool? Bulk { get; set; }
+
         /// <summary>
         /// Date the subscription started.
         /// </summary>
@@ -106,11 +108,15 @@ namespace Recurly
         /// Date the trial started, if the subscription has a trial.
         /// </summary>
         public DateTime? TrialPeriodStartedAt { get; private set; }
+        /// <summary>
+        /// Date the Bank Account has been authorized for this subscription
+        /// </summary>
+        public DateTime? BankAccountAuthorizedAt { get; set; }
 
         /// <summary>
         /// Date the trial ends, if the subscription has/had a trial.
-        /// 
-        /// This may optionally be set on new subscriptions to specify an exact time for the 
+        ///
+        /// This may optionally be set on new subscriptions to specify an exact time for the
         /// subscription to commence.  The subscription will be active and in "trial" until
         /// this date.
         /// </summary>
@@ -130,7 +136,7 @@ namespace Recurly
         private DateTime? _trialPeriodEndsAt;
 
         /// <summary>
-        /// If set, the subscription will begin in the future on this date. 
+        /// If set, the subscription will begin in the future on this date.
         /// The subscription will apply the setup fee and trial period, unless the plan has no trial.
         /// </summary>
         public DateTime? StartsAt { get; set; }
@@ -338,10 +344,16 @@ namespace Recurly
             Preview(ChangeTimeframe.Now);
         }
 
-        public void Postpone(DateTime nextRenewalDate)
+        /// <summary>
+        /// For an active subscription, this will pause the subscription until the specified date.
+        /// </summary>
+        /// <param name="nextRenewalDate">The specified time the subscription will be postponed</param>
+        /// <param name="bulk">bulk = false (default) or true to bypass the 60 second wait while postponing</param>
+        public void Postpone(DateTime nextRenewalDate, bool bulk = false)
+
         {
             Client.Instance.PerformRequest(Client.HttpRequestMethod.Put,
-                UrlPrefix + Uri.EscapeUriString(Uuid) + "/postpone?next_renewal_date=" + nextRenewalDate.ToString("yyyy-MM-ddThh:mm:ssZ"),
+                UrlPrefix + Uri.EscapeUriString(Uuid) + "/postpone?next_renewal_date=" + nextRenewalDate.ToString("yyyy-MM-ddThh:mm:ssZ") + "&bulk=" + bulk.ToString().ToLower(),
                 ReadXml);
         }
 
@@ -446,7 +458,7 @@ namespace Recurly
                         if (DateTime.TryParse(reader.ReadElementContentAsString(), out dateVal))
                             CurrentPeriodStartedAt = dateVal;
                         break;
-                        
+
                     case "current_period_ends_at":
                         if (DateTime.TryParse(reader.ReadElementContentAsString(), out dateVal))
                             CurrentPeriodEndsAt = dateVal;
@@ -460,6 +472,11 @@ namespace Recurly
                     case "trial_ends_at":
                         if (DateTime.TryParse(reader.ReadElementContentAsString(), out dateVal))
                             _trialPeriodEndsAt = dateVal;
+                        break;
+
+                    case "bank_account_authorized_at":
+                        if (DateTime.TryParse(reader.ReadElementContentAsString(), out dateVal))
+                            BankAccountAuthorizedAt = dateVal;
                         break;
 
                     case "subscription_add_ons":
@@ -582,6 +599,9 @@ namespace Recurly
             if (TrialPeriodEndsAt.HasValue)
                 xmlWriter.WriteElementString("trial_ends_at", TrialPeriodEndsAt.Value.ToString("s"));
 
+            if (BankAccountAuthorizedAt.HasValue)
+                xmlWriter.WriteElementString("bank_account_authorized_at", BankAccountAuthorizedAt.Value.ToString("s"));
+
             if (StartsAt.HasValue)
                 xmlWriter.WriteElementString("starts_at", StartsAt.Value.ToString("s"));
 
@@ -590,6 +610,9 @@ namespace Recurly
 
             if (FirstRenewalDate.HasValue)
                 xmlWriter.WriteElementString("first_renewal_date", FirstRenewalDate.Value.ToString("s"));
+
+            if (Bulk.HasValue)
+                xmlWriter.WriteElementString("bulk", Bulk.ToString().ToLower());
 
             if (CollectionMethod.Like("manual"))
             {
@@ -688,7 +711,7 @@ namespace Recurly
     {
         /// <summary>
         /// Returns a list of recurly subscriptions
-        /// 
+        ///
         /// A subscription will belong to more than one state.
         /// </summary>
         /// <param name="state">State of subscriptions to return, defaults to "live"</param>